fix: Session 139 mypy fixes - 235 errors (50 files)
Progress: 6678 → 6443 errors (3.5% reduction)
10 parallel coordinators working on domain-specific files.

+++ backend/app/multi_objective/decision_support.py
@@ -496,7 +496,7 @@ class SolutionExplorer:
 
             dist += (val1 - val2) ** 2
 
-        return np.sqrt(dist)
+        return float(np.sqrt(dist))
 
     def _count_remaining(
         self,
